consolidate streaming code in TentD::Streaming
